WebhookList now checks for private channels, public modifiers removed, methods collapsed
1. WebhookList checks for IPrivateChannel and throws
UnsupportedOperationException
2. Remove unnecessary public modifiers on the fields in IMessage
3. Collapse the execute methods down to message, message + everything,
file, file + everything, input stream, input stream + everything,
embeds, embeds + everything
4. Extracted IEmbedded into its own file, and EmbedObject into its own
file as well.
