+++ src/main/java/sx/blah/discord/api/internal/json/requests/WebhookExecuteRequest.java
@@ -1,7 +1,8 @@
 package sx.blah.discord.api.internal.json.requests;
 
+import sx.blah.discord.api.internal.json.objects.EmbedObject;
 import sx.blah.discord.api.internal.json.objects.MessageObject;
-import sx.blah.discord.handle.obj.IMessage;
+import sx.blah.discord.handle.obj.IEmbedded;
 
 /**
  * Represents a request to execute a webhook
@@ -31,7 +32,7 @@ public class WebhookExecuteRequest {
 	/**
 	 * An array of embedded rich content
 	 */
-	public MessageObject.EmbedObject[] embeds;
+	public EmbedObject[] embeds;
 
 	public WebhookExecuteRequest(String content, String username, String avatar_url, boolean tts) {
 		this.content = content;
@@ -40,18 +41,18 @@ public class WebhookExecuteRequest {
 		this.tts = tts;
 	}
 
-	public WebhookExecuteRequest(String username, String avatar_url, MessageObject.EmbedObject[] embeds) {
+	public WebhookExecuteRequest(String username, String avatar_url, EmbedObject[] embeds) {
 		this.username = username;
 		this.avatar_url = avatar_url;
 		this.embeds = embeds;
 	}
 
-	public WebhookExecuteRequest(String username, String avatar_url, IMessage.IEmbedded[] embeds) {
+	public WebhookExecuteRequest(String username, String avatar_url, IEmbedded[] embeds) {
 		this.username = username;
 		this.avatar_url = avatar_url;
-		this.embeds = new MessageObject.EmbedObject[embeds.length];
+		this.embeds = new EmbedObject[embeds.length];
 		for (int i = 0; i < embeds.length; i++) {
-			this.embeds[i] = new MessageObject.EmbedObject(embeds[i]);
+			this.embeds[i] = new EmbedObject(embeds[i]);
 		}
 	}
 }
